The STID Class

The STID (Integrated Single-cell Transcriptomics) object extends the Seurat class to add specialized functionality for analyzing infection and spatial transcriptomics data. It inherits all functionality from Seurat objects while adding additional slots for tracking infection-related analyses.

Slots

STID_analysis

An STID_analysis object containing all infection-specific analysis results and metadata

assays

A list of assays for this project (inherited from Seurat)

meta.data

Contains meta-information about each cell (inherited from Seurat)

active.assay

Name of the active, or default, assay (inherited from Seurat)

active.ident

The active cluster identity (inherited from Seurat)

graphs

A list of Graph objects (inherited from Seurat)

neighbors

Neighbor graphs (inherited from Seurat)

reductions

A list of dimensional reduction objects (inherited from Seurat)

images

A list of spatial image objects (inherited from Seurat)

project.name

Name of the project (inherited from Seurat)

misc

A list of miscellaneous information (inherited from Seurat)

version

Version of Seurat this object was built under (inherited from Seurat)

commands

A list of logged commands (inherited from Seurat)

tools

A list of miscellaneous data generated by other tools (inherited from Seurat)

See also

Examples

if (FALSE) { # \dontrun{
# Create an STID object from a Seurat object
ist_obj <- as.STID(
  seurat_obj = seurat_object,
  samp_colnm = "sample_id",
  celltype_colnm = "cell_type",
  host_org = "human",
  pathogen_grp = "virus",
  pathogen_org = "SARS-CoV-2",
  data_format = "StereoSeq"
)
} # }